`core` submodule
