+++ rollup.config.mjs
@@ -10,6 +10,7 @@ import typescript from 'typescript';
 export default defineConfig({
   input: {
     'context-values': './src/index.ts',
+    'context-values.core': './src/core/index.ts',
     'context-values.updatable': './src/updatable/index.ts',
   },
   plugins: [
@@ -23,6 +24,9 @@ export default defineConfig({
   ],
   external: externalModules(),
   manualChunks(id) {
+    if (id.startsWith(path.resolve('src', 'core') + path.sep)) {
+      return 'context-values.core';
+    }
     if (id.startsWith(path.resolve('src', 'updatable') + path.sep)) {
       return 'context-values.updatable';
     }
@@ -43,6 +47,12 @@ export default defineConfig({
           declarationMap: true,
         },
         entries: {
+          core: {
+            file: 'core/index.d.ts',
+            refs: false, // Do not refer to `../index.d.ts`.
+                         // Such reference breaks `rollup-plugin-typescript2`
+                         // within pnpm workspace installation.
+          },
           updatable: {
             file: 'updatable/index.d.ts',
             refs: false, // Do not refer to `../index.d.ts`.
