+++ src/core/assets/index.ts
@@ -1,0 +1,3 @@
+export * from './alias.asset';
+export * from './build.asset';
+export * from './const.asset';
