+++ src/core/static/array.entry.ts
