+++ src/core/static/index.ts
@@ -1,0 +1,2 @@
+export * from './array.entry';
+export * from './single.entry';
