+++ src/core/static/single.entry.ts
