+++ src/index.ts
@@ -2,13 +2,13 @@
  * @packageDocumentation
  * @module Module @proc7ts/context-values
  */
+export * from './build';
 export * from './context-key-error';
 export * from './context-request';
 export * from './context-ref';
 export * from './context-values';
 export * from './conventional';
-export * from './entry';
+export * from './core';
 export * from './registry';
 export * from './singleton';
-export * from './static';
 export * from './key';
