--- src/static/index.ts
@@ -1,2 +1,0 @@
-export * from './array';
-export * from './single';
